Make prefix stripping reusable
